# Refactor parsing logic of the Query Builder (second try)