… src/main/java/org/opensearch/knn/index/query/parser/KNNQueryBuilderParser.java
… src/test/java/org/opensearch/knn/index/query/KNNQueryBuilderTests.java
… src/test/java/org/opensearch/knn/index/query/parser/KNNQueryParserTests.java