@shatejas 2024-07-29
> So in neural, we cannot directly re-use KNNQueryBuilder IO because there are different required parameters

I see, might be possible if we club all knn related fields together, but again we will have to check if that can be done in backward compatible way

> Also, why did you define NPROBE in KNNQueryBuilder then?

All fields were defined there, didn't think much back then